move data retention to its own file
